minor: webpack config + dep
